+++ src/main/java/com/bsuir/masasha/hostel/web/controller/admin/AdminHotelEditController.java
@@ -55,7 +55,7 @@ public class AdminHotelEditController {
     }
 
     @PostMapping("/add/roomType")
-    public String addRoomType(@RequestParam String roomTypeTitle, @RequestParam Double roomTypeCost) {
+    public String addRoomType(@RequestParam String roomTypeTitle, @RequestParam Integer roomTypeCost) {
         hotelEditService.addRoomType(roomTypeTitle, roomTypeCost);
 
         return REDIRECT + HOTEL_EDITOR_MAPPING;
